+++ src/Infrastructure/CosmosCity.cs
@@ -7,9 +7,9 @@ namespace Infrastructure
 {
     public class CosmosCity : BaseEntity
     {
-        public int ZombieCount { get; set; }
-        public int KangarooCount { get; set; }
-        public int HumanCount { get; set; }
+        public Int64 ZombieCount { get; set; }
+        public Int64 KangarooCount { get; set; }
+        public Int64 HumanCount { get; set; }
         public string Name { get { return Id; } set { Id = value; } }
         public string State { get; set; }
